Getting unit test db to work without template
