+++ src/fbo_scraper/db/db.py
@@ -1,6 +1,7 @@
 import datetime
 
-from sqlalchemy.ext.declarative import declarative_base
+#from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import declarative_base
 from sqlalchemy import (
     Column,
     Integer,
